Moved more code out of main
